refactor(local-inference): consolidate types in @elizaos/shared, drop dead UI providers runtime
Eliminates the ~300-line type-definition duplication between
`packages/ui/src/services/local-inference/types.ts` and
`packages/app-core/src/services/local-inference/types.ts`. After
comment-stripping the two files were byte-identical (both 216 lines
of actual TypeScript), so the "server-side superset" framing in the
old header comments was wrong — the same exports lived in both
places. Now both files are 1-line re-exports from the shared module.

`@elizaos/shared/local-inference` now owns:
- AgentModelSlot / TextGenerationSlot / AGENT_MODEL_SLOTS /
  TEXT_GENERATION_SLOTS / ModelAssignments / InstalledModel
  (already there before this PR)
- ModelBucket / ModelCategory / TokenizerFamily
- LocalRuntimeBackend / LocalRuntimeKernel / LocalRuntimeOptimizations
  / LocalRuntimeAcceleration
- CatalogModel / HardwareProbe / MobileHardwareProbe / HardwareFitLevel
- DownloadState / DownloadJob / LocalInferenceDownloadStatus
  / DownloadEvent / ActiveModelState
- LocalInferenceSlotReadiness / LocalInferenceReadiness / ModelHubSnapshot
- ProviderId / ProviderEnableState / ProviderDefinition / ProviderStatus
  (new providers-types.ts file)

Side cleanup — UI providers.ts runtime deleted:
- `packages/ui/src/services/local-inference/providers.ts` was 476 lines
  of `BUILT_IN_PROVIDERS`, `snapshotProviders()`, per-provider
  `getEnableState` readers — all dead at UI runtime. The only UI
  consumers (`ios-local-agent-kernel.ts`, `client-local-inference.ts`)
  imported `import type { ProviderStatus }` only.
- The authoritative `/api/local-inference/providers` response comes
  from `packages/app-core/src/api/local-inference-compat-routes.ts`
  which imports `snapshotProviders` from the app-core copy. The iOS
  local-agent kernel ships its own one-provider response via
  `capacitorLlamaProviderStatus()` and never touches BUILT_IN_PROVIDERS.
- The UI runtime had also drifted from the server: it lacked the
  account-pool integration, mis-declared TEXT_EMBEDDING support for
  Anthropic, and stubbed `apiKeyOrLinkedAccountState`. Deleting it
  removes a footgun.

Both `import type { ProviderStatus }` callsites switched to
`@elizaos/shared`.

Verification:
- 503 schema/contract tests pass in @elizaos/shared
- app-core + ui typecheck show zero new errors; remaining failures
  are pre-existing on develop (capacitor-bridge / plugin-workflow
  / plugin-streaming module-resolution issues unrelated to this PR)
