+++ packages/shared/src/local-inference/index.ts
@@ -3,11 +3,10 @@
  *
  * Shared local-inference contract used by both the server-side service
  * (`@elizaos/app-core/src/services/local-inference`) and the UI client
- * (`@elizaos/ui/src/services/local-inference`). Only modules that are
- * byte-identical between the two and that share the exact same semantics
- * live here. Server-only logic (KV cache management, llama-server
- * lifecycle, conversation registry, metrics scraping) stays in
- * `app-core`.
+ * (`@elizaos/ui/src/services/local-inference`). Type definitions live
+ * here; runtime logic stays in `app-core` (server-side KV cache
+ * management, llama-server lifecycle, conversation registry, metrics)
+ * and `ui` (client wiring against the agent API).
  */
 
 export {
@@ -29,6 +28,12 @@ export {
   localInferenceRoot,
   registryPath,
 } from "./paths.js";
+export type {
+  ProviderDefinition,
+  ProviderEnableState,
+  ProviderId,
+  ProviderStatus,
+} from "./providers-types.js";
 export {
   DEFAULT_ROUTING_POLICY,
   type RoutingPolicy,
@@ -39,11 +44,31 @@ export {
   writeRoutingPreferences,
 } from "./routing-preferences.js";
 export {
+  type ActiveModelState,
   AGENT_MODEL_SLOTS,
   type AgentModelSlot,
+  type CatalogModel,
+  type DownloadEvent,
+  type DownloadJob,
+  type DownloadState,
+  type HardwareFitLevel,
+  type HardwareProbe,
   type InstalledModel,
+  type LocalInferenceDownloadStatus,
+  type LocalInferenceReadiness,
+  type LocalInferenceSlotReadiness,
+  type LocalRuntimeAcceleration,
+  type LocalRuntimeBackend,
+  type LocalRuntimeKernel,
+  type LocalRuntimeOptimizations,
+  type MobileHardwareProbe,
   type ModelAssignments,
+  type ModelBucket,
+  type ModelCategory,
+  type ModelHubSnapshot,
+  TEXT_GENERATION_SLOTS,
   type TextGenerationSlot,
+  type TokenizerFamily,
 } from "./types.js";
 export {
   __registryPathForTests,
